chore(gohai): properly rename processes files
