+++ pkg/gohai/processes/processes.go
@@ -3,45 +3,41 @@
 // Copyright © 2015 Kentaro Kuribayashi <kentarok@gmail.com>
 // Copyright 2014-present Datadog, Inc.
 
-//go:build linux || darwin
-// +build linux darwin
-
-// Package processes regroups collecting information about existing processes
+// Package processes regroups collecting information about running processes.
 package processes
 
-import (
-	"strings"
-	"time"
+import "flag"
 
-	"github.com/DataDog/datadog-agent/pkg/gohai/processes/gops"
-)
+var options struct {
+	limit int
+}
 
-// ProcessField is an untyped representation of a process group,
-// compatible with the legacy "processes" resource check.
-type ProcessField [7]interface{}
+// Processes is the Collector type of the processes package.
+type Processes struct{}
 
-// getProcesses return a JSON payload which is compatible with
-// the legacy "processes" resource check
-func getProcesses(limit int) ([]interface{}, error) {
-	processGroups, err := gops.TopRSSProcessGroups(limit)
-	if err != nil {
+const name = "processes"
+
+func init() {
+	flag.IntVar(&options.limit, name+"-limit", 20, "Number of process groups to return")
+}
+
+// Name returns the name of the package
+func (processes *Processes) Name() string {
+	return name
+}
+
+// Collect collects the processes information.
+// Returns an object which can be converted to a JSON or an error if nothing could be collected.
+// Tries to collect as much information as possible.
+func (processes *Processes) Collect() (result interface{}, err error) {
+	// even if getProcesses returns nil, simply assigning to result
+	// will have a non-nil return, because it has a valid inner
+	// type (more info here: https://golang.org/doc/faq#nil_error )
+	// so, jump through the hoop of temporarily storing the return,
+	// and explicitly return nil if it fails.
+	gpresult, err := getProcesses(options.limit)
+	if gpresult == nil {
 		return nil, err
 	}
-
-	snapData := make([]ProcessField, len(processGroups))
-
-	for i, processGroup := range processGroups {
-		processField := ProcessField{
-			strings.Join(processGroup.Usernames(), ","),
-			0, // pct_cpu, requires two consecutive samples to be computed, so not fetched for now
-			processGroup.PctMem(),
-			processGroup.VMS(),
-			processGroup.RSS(),
-			processGroup.Name(),
-			len(processGroup.Pids()),
-		}
-		snapData[i] = processField
-	}
-
-	return []interface{}{time.Now().Unix(), snapData}, nil
+	return gpresult, err
 }
